Char becomes Letter
It was all ok from the beggin', we just needed to set the array itself
in the start function too. Who'd know, right? Anyway, adjust the image
size, so the game don't crash, a smaller spawn rate IF you lose the
letter, and the Intent stuff, that is the whole purpose of this work.
Try not to forget about the performance, kay? Good work, tho